requirements.txt
Django==4.2.14
django-bootstrap5==24.2
django-environ==0.11.2
django-import-export==4.1.1
django-pandas==0.6.7
plotly==5.13.1
